Allow static plugins
Developers can now pass the STATIC argument that will build plugins
as static libraries with the appropriate Qt defines.

Static plugins are not installed, to do this we have to generate
CMake code to allow external project to use the plugins.
For now we just want to import static plugins within their
project.

Add missing QT_PLUGIN define while we are here.

Closes: #7