LiriAddModule: fix private headers forwarding
Always create the forward header, some headers might be generated
later like Wayland protocol headers.

Prepend the current source directory only to relative file names,
we now expect the user to give us an absolute path when it needs to
specify a file name not in the current source directory.