LiriAddModule: automatically figure out the class names
When FORWARDING_HEADERS is not specified, we now detect the class names
automatically.

Also fix header generation, local headers were not forwarded.

Since class name autodetection is based off of a piece of code from
syncqt.pl (rewritten in Python and improved), we change the license of
the whole project to GPL 3.0 just to be sure.

Closes: #15