Revamp how targets are created
Sometimes we want to extend the targets by adding more sources, headers,
resources, so on and so forth.

However we build those right after the target has been declared, but the
target is extended only later when it's too late.

Split the target declaration in more steps: first we add the target, then
we optionally extend it and with complete its setup.

Class name headers are always automatically detected, but users can add
to the list with CLASS_HEADERS (instead of FORWARDING_HEADERS).